Moved numerical operations test to a single file.
